+++ edward/inferences/conjugacy/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from edward.inferences.conjugacy.conjugacy import *
